Update style.css
Changed formatting for edit and delete buttons. Edit is now yellow, and delete is now red. Also added btn-std formatting for buttons with default properties (standard blue in our case). Changed formatting for the table to hide the ID field, so it can be returned without being displayed on the page.